requirements.txt
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
cloudpickle==2.1.0
cycler==0.11.0
dateparser==1.1.1
decorator==5.1.1
executing==0.10.0
express==0.1.5
ffmpeg==1.4
fonttools==4.36.0
gym==0.21.0
gym-notices==0.0.8
importlib-metadata==4.12.0
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
matplotlib-scalebar==0.8.1
neo==0.10.2
numpy==1.23.2
packaging==21.3
pandas==1.4.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
quantities==0.13.0
regex==2022.3.2
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
stable-baselines3==1.6.0
stack-data==0.4.0
tenacity==8.0.1
threadpoolctl==3.1.0
torch==1.12.1
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
wcwidth==0.2.5
zipp==3.8.1
